Remove .specification from .gitignore
With .specification in .gitignore, every developer on a Rails project
will see the dreaded 'no specification file' warning in their local clone
of the repo.  When a developer runs 'rake gems:refresh_specs' to fix it
locally, his changes never get committed to Git.  So every developer on
the project gets the enjoyment of generating the .specification file.